propagate the Optional parameters from base class for typing. fix test mock path
