+++ mypy.ini
@@ -61,6 +61,7 @@ files = src/sentry/analytics/,
         src/sentry/integrations/slack/,
         src/sentry/integrations/message_builder.py,
         src/sentry/integrations/vsts/,
+        src/sentry/issues/search.py,
         src/sentry/killswitches.py,
         src/sentry/lang/native/appconnect.py,
         src/sentry/mail/notifications.py,
